requirements.txt
aiomysql==0.0.21
discord.py==1.7.3
